requirements.txt
-i https://pypi.org/simple

asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
decorator==5.1.1; python_version >= '3.7'
executing==0.10.0
faker==14.1.0
idna==3.3; python_version >= '3.5'
ipdb==0.13.9
ipython==8.4.0; python_version >= '3.7'
jedi==0.18.1; python_version >= '3.6'
jinja2==3.1.2
markupsafe==2.1.1; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
parso==0.8.3; python_version >= '3.6'
peewee==3.15.1
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pid==3.0.4
prompt-toolkit==3.0.30; python_full_version >= '3.6.2'
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.13.0; python_version >= '3.6'
pytelegrambotapi==4.7.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
requests==2.28.1; python_version >= '3.7' and python_version < '4'
setuptools==65.2.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.4.0
tenacity==8.0.1
toml==0.10.2; python_version >= '3.7'
traitlets==5.3.0; python_version >= '3.7'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
wcwidth==0.2.5
